refactor: break up into multiple files
This commit also introduces a new XMLDeserialization type that hopefully will make it easier to implement deserialization for custom types.
