+++ Source/Double+XMLDeserialization.swift
@@ -1,0 +1,61 @@
+//
+//  XMLDeserializationError.swift
+//  SWXMLHash
+//
+//  Copyright (c) 2016 Maciek Grzybowskio, 2022 David Mohundro
+//
+//  Permission is hereby granted, free of charge, to any person obtaining a copy
+//  of this software and associated documentation files (the "Software"), to deal
+//  in the Software without restriction, including without limitation the rights
+//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//  copies of the Software, and to permit persons to whom the Software is
+//  furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in
+//  all copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+//  THE SOFTWARE.
+//
+
+import Foundation
+
+extension Double: XMLDeserialization, XMLElementDeserializable {
+    /**
+    Attempts to deserialize XML element content to a Double
+
+    - parameters:
+        - element: the XMLElement to be deserialized
+    - throws: an XMLDeserializationError.typeConversionFailed if the element cannot be deserialized
+    - returns: the deserialized Double value
+    */
+    public static func deserialize(_ element: XMLElement) throws -> Double {
+        guard let value = Double(try element.nonEmptyTextOrThrow()) else {
+            throw XMLDeserializationError.typeConversionFailed(type: "Double", element: element)
+        }
+        return value
+    }
+
+    /**
+     Attempts to deserialize XML attribute content to a Double
+
+     - parameter attribute: The XMLAttribute to be deserialized
+     - throws: an XMLDeserializationError.attributeDeserializationFailed if the attribute cannot be
+               deserialized
+     - returns: the deserialized Double value
+     */
+    public static func deserialize(_ attribute: XMLAttribute) throws -> Double {
+        guard let value = Double(attribute.text) else {
+            throw XMLDeserializationError.attributeDeserializationFailed(
+                    type: "Double", attribute: attribute)
+        }
+        return value
+    }
+
+    public func validate() {}
+}
