+++ Source/Int+XMLDeserialization.swift
@@ -1,0 +1,61 @@
+//
+//  XMLDeserializationError.swift
+//  SWXMLHash
+//
+//  Copyright (c) 2016 Maciek Grzybowskio, 2022 David Mohundro
+//
+//  Permission is hereby granted, free of charge, to any person obtaining a copy
+//  of this software and associated documentation files (the "Software"), to deal
+//  in the Software without restriction, including without limitation the rights
+//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//  copies of the Software, and to permit persons to whom the Software is
+//  furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in
+//  all copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+//  THE SOFTWARE.
+//
+
+import Foundation
+
+extension Int: XMLDeserialization {
+    /**
+    Attempts to deserialize XML element content to a Int
+
+    - parameters:
+        - element: the XMLElement to be deserialized
+    - throws: an XMLDeserializationError.typeConversionFailed if the element cannot be deserialized
+    - returns: the deserialized Int value
+    */
+    public static func deserialize(_ element: XMLElement) throws -> Int {
+        guard let value = Int(try element.nonEmptyTextOrThrow()) else {
+            throw XMLDeserializationError.typeConversionFailed(type: "Int", element: element)
+        }
+        return value
+    }
+
+    /**
+     Attempts to deserialize XML attribute content to an Int
+
+     - parameter attribute: The XMLAttribute to be deserialized
+     - throws: an XMLDeserializationError.attributeDeserializationFailed if the attribute cannot be
+               deserialized
+     - returns: the deserialized Int value
+     */
+    public static func deserialize(_ attribute: XMLAttribute) throws -> Int {
+        guard let value = Int(attribute.text) else {
+            throw XMLDeserializationError.attributeDeserializationFailed(
+                    type: "Int", attribute: attribute)
+        }
+        return value
+    }
+
+    public func validate() {}
+}
