+++ Source/XMLIndexerDeserializable.swift
@@ -1,0 +1,56 @@
+//
+//  XMLDeserializationError.swift
+//  SWXMLHash
+//
+//  Copyright (c) 2016 Maciek Grzybowskio, 2022 David Mohundro
+//
+//  Permission is hereby granted, free of charge, to any person obtaining a copy
+//  of this software and associated documentation files (the "Software"), to deal
+//  in the Software without restriction, including without limitation the rights
+//  to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+//  copies of the Software, and to permit persons to whom the Software is
+//  furnished to do so, subject to the following conditions:
+//
+//  The above copyright notice and this permission notice shall be included in
+//  all copies or substantial portions of the Software.
+//
+//  THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+//  IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+//  FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+//  AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+//  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+//  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+//  THE SOFTWARE.
+//
+
+import Foundation
+
+/// Provides XMLIndexer deserialization / type transformation support
+public protocol XMLIndexerDeserializable {
+    /// Method for deserializing elements from XMLIndexer
+    static func deserialize(_ element: XMLIndexer) throws -> Self
+    /// Method for validating elements post deserialization
+    func validate() throws
+}
+
+/// Provides XMLIndexer deserialization / type transformation support
+public extension XMLIndexerDeserializable {
+    /**
+    A default implementation that will throw an error if it is called
+
+    - parameters:
+        - element: the XMLIndexer to be deserialized
+    - throws: an XMLDeserializationError.implementationIsMissing if no implementation is found
+    - returns: this won't ever return because of the error being thrown
+    */
+    static func deserialize(_ element: XMLIndexer) throws -> Self {
+        throw XMLDeserializationError.implementationIsMissing(
+                method: "XMLIndexerDeserializable.deserialize(element: XMLIndexer)")
+    }
+
+    /**
+    A default do nothing implementation of validation.
+    - throws: nothing
+    */
+    func validate() throws {}
+}
